Time: 5 ms (20.20%) | Memory: 8.5 MB (84.02%) - LeetSync

+++ 7-reverse-integer/reverse-integer.cpp
@@ -2,11 +2,11 @@ class Solution {
 public:
     int reverse(int n) {
 
-        long rev = 0 ;
+        int rev = 0 ;
 
         while( n != 0)
         {
-            long rem = n % 10 ;
+            int rem = n % 10 ;
 
             if( rev > INT_MAX / 10 || rev < INT_MIN / 10) return 0;
